Move tests to top-level namespace
