+++ tests/Alley/ReshapeTest.php
@@ -9,11 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Alley\Tests\Reshape;
+namespace Alley;
 
 use PHPUnit\Framework\TestCase;
-
-use function Alley\reshape;
 
 final class ReshapeTest extends TestCase
 {
